+++ docs/changes.rst
@@ -6,6 +6,7 @@ Release history
 - feat: new event type "businessTexts"
 - feat: :class:`wxc_sdk.licenses.License` attributes: consumed_by_users, consumed_by_workspaces
 - feat: :meth:`wxc_sdk.person_settings.voicemail.VoicemailApi.modify_passcode` to set voicemail passcode for users
+- feat: guests API :attr:`wxc_sdk.WebexSimpleApi.guests`
 
 1.17.1
 ------